♻️ refactor: 마이페이지 파일 구조 변경
